requirements.txt
fastapi==0.115.0
httpx==0.27.2
numpy==2.1.1
pgvector==0.3.2
pydantic==2.9.2
pydantic_settings==2.5.2
python-dotenv==1.0.1
Requests==2.32.3
SQLAlchemy==2.0.34
uvicorn==0.30.6
